Typescript rewrite
