+++ .eslintrc.js
@@ -1,0 +1,11 @@
+module.exports = {
+  root: true,
+  parser: '@typescript-eslint/parser',
+  plugins: ['@typescript-eslint'],
+  extends: ['eslint:recommended', 'plugin:@typescript-eslint/recommended'],
+
+  rules: {
+    // allow while(true) loops
+    'no-constant-condition': ['error', { checkLoops: false }],
+  },
+}
