+++ __tests__/global.d.ts
@@ -1,0 +1,1 @@
+import 'jest-extended'
