+++ __tests__/pgn/2.pgn
@@ -1,0 +1,2 @@
+1. f3 e5
+2. g4 Qh4#
