+++ __tests__/pgn/3.pgn
@@ -1,0 +1,5 @@
+[SetUp "1"]
+[FEN "r1bqk1nr/pppp1ppp/2n5/4p3/1bB1P3/2P2N2/P2P1PPP/RNBQK2R b KQkq - 0 1"]
+
+1. ... Ba5 2. O-O d6
+3. d4
